+++ CHANGES.rst
@@ -12,6 +12,8 @@ master (in development)
 
   - Automatically disable syntax checkers that report too many errors [GH-476]
   - Reduce filesystem access when parsing errors to improve parsing speed
+  - Add explicit `load-path` inheritance to `flycheck-emacs-lisp-load-path`, via
+    new `inherit` value [GH-511]
 
 - Bug fixes: